refactoring python import by changing folders hierarchy
