+++ main.tf
@@ -181,7 +181,7 @@ resource "aws_lambda_function" "this" {
   filename         = data.archive_file.this.output_path
   function_name    = var.name
   role             = var.custom_iam_role_arn == null ? aws_iam_role.this[0].arn : var.custom_iam_role_arn
-  handler          = "main.lambda_handler"
+  handler          = "scheduler.main.lambda_handler"
   source_code_hash = data.archive_file.this.output_base64sha256
   runtime          = "python3.7"
   timeout          = "600"
