+++ tox.ini
@@ -23,7 +23,7 @@ skip_install = true
 deps =
     black>=19.3b0
 commands =
-    black package/ --line-length 78 --check
+    black package/ --line-length 79 --check
 
 # Linters
 [testenv:flake8]
